NLP - Sarcasm Detection
Past studies in Sarcasm Detection mostly make use of Twitter datasets collected using hashtag based supervision but such 
datasets are noisy in terms of labels and language. Furthermore, many tweets are replies to other tweets and detecting 
sarcasm in these requires the availability of contextual tweets. 
 
In this hands-on project, the goal is to build a model to detect whether a sentence is sarcastic or not, using Bidirectional 
LSTMs.